Create clients and components
