+++ src/client/components/Login.tsx
@@ -2,7 +2,9 @@ import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 const Login: React.FC = () => {
-  return <div></div>;
+  return <div>
+    
+  </div>;
 };
 
 export default Login;
