+++ src/client/components/Maps.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 const Maps: React.FC = () => {
-  return <div></div>;
+  return <div>'hi babe'</div>;
 };
 
 export default Maps;
