+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
+import App from './client/components/App';
 
 const root = createRoot(document.getElementById('root'));
 root.render(
